+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.8.3
  - Fix missing `metadata` and `type` of the last event [#223](https://github.com/logstash-plugins/logstash-input-s3/pull/223)
+ - Not stopping processing in the middle of a file when a shutdown was initiated
 
 ## 3.8.2
  - Refactor: read sincedb time once per bucket listing [#233](https://github.com/logstash-plugins/logstash-input-s3/pull/233)